Implement basic Dropbox integration (--dropbox).
When generating KML, the tool will lookup image in designated Dropbox
folder, share it, and use Dropbox-based links to images. This means
That it will be possible to pass just KML file and another party will
be able to use it with Google Earth. Or, KML can be imported into Google
Maps and images will automagically work either.